requirements.txt
flask==3.1.0
requests==2.32.5
pandas==2.3.3
python-dotenv==1.2.2
